moved gps code and cleaned it up
